+++ lib/providers/problem.dart
@@ -9,6 +9,7 @@ import 'package:miel_work_app/models/organization.dart';
 import 'package:miel_work_app/models/problem.dart';
 import 'package:miel_work_app/models/user.dart';
 import 'package:miel_work_app/services/fm.dart';
+import 'package:miel_work_app/services/log.dart';
 import 'package:miel_work_app/services/problem.dart';
 import 'package:miel_work_app/services/user.dart';
 import 'package:path/path.dart' as p;
@@ -17,6 +18,7 @@ class ProblemProvider with ChangeNotifier {
   final ProblemService _problemService = ProblemService();
   final UserService _userService = UserService();
   final FmService _fmService = FmService();
+  final LogService _logService = LogService();
 
   Future<String?> create({
     required OrganizationModel? organization,
@@ -121,6 +123,17 @@ class ProblemProvider with ChangeNotifier {
           }
         }
       }
+      //ログ保存
+      String logId = _logService.id();
+      _logService.create({
+        'id': logId,
+        'organizationId': organization.id,
+        'content': 'クレーム／要望を追加しました。',
+        'device': 'SP(アプリ)',
+        'createdUserId': loginUser.id,
+        'createdUserName': loginUser.name,
+        'createdAt': DateTime.now(),
+      });
     } catch (e) {
       error = 'クレーム／要望の追加に失敗しました';
     }
@@ -223,6 +236,17 @@ class ProblemProvider with ChangeNotifier {
           'count': count,
         });
       }
+      //ログ保存
+      String logId = _logService.id();
+      _logService.create({
+        'id': logId,
+        'organizationId': organization.id,
+        'content': 'クレーム／要望情報を編集しました。',
+        'device': 'SP(アプリ)',
+        'createdUserId': loginUser.id,
+        'createdUserName': loginUser.name,
+        'createdAt': DateTime.now(),
+      });
     } catch (e) {
       error = 'クレーム／要望情報の編集に失敗しました';
     }
@@ -275,6 +299,17 @@ class ProblemProvider with ChangeNotifier {
           }
         }
       }
+      //ログ保存
+      String logId = _logService.id();
+      _logService.create({
+        'id': logId,
+        'organizationId': organization.id,
+        'content': 'クレーム／要望に社内コメントを追記しました。',
+        'device': 'SP(アプリ)',
+        'createdUserId': loginUser.id,
+        'createdUserName': loginUser.name,
+        'createdAt': DateTime.now(),
+      });
     } catch (e) {
       error = '社内コメントの追記に失敗しました';
     }
@@ -283,12 +318,25 @@ class ProblemProvider with ChangeNotifier {
 
   Future<String?> process({
     required ProblemModel problem,
+    required UserModel? loginUser,
   }) async {
     String? error;
+    if (loginUser == null) return 'クレーム／要望情報の処理に失敗しました';
     try {
       _problemService.update({
         'id': problem.id,
         'processed': true,
+      });
+      //ログ保存
+      String logId = _logService.id();
+      _logService.create({
+        'id': logId,
+        'organizationId': problem.organizationId,
+        'content': 'クレーム／要望を処理済しました。',
+        'device': 'SP(アプリ)',
+        'createdUserId': loginUser.id,
+        'createdUserName': loginUser.name,
+        'createdAt': DateTime.now(),
       });
     } catch (e) {
       error = 'クレーム／要望情報の処理に失敗しました';
@@ -298,11 +346,24 @@ class ProblemProvider with ChangeNotifier {
 
   Future<String?> delete({
     required ProblemModel problem,
+    required UserModel? loginUser,
   }) async {
     String? error;
+    if (loginUser == null) return 'クレーム／要望情報の削除に失敗しました';
     try {
       _problemService.delete({
         'id': problem.id,
+      });
+      //ログ保存
+      String logId = _logService.id();
+      _logService.create({
+        'id': logId,
+        'organizationId': problem.organizationId,
+        'content': 'クレーム／要望を削除しました。',
+        'device': 'SP(アプリ)',
+        'createdUserId': loginUser.id,
+        'createdUserName': loginUser.name,
+        'createdAt': DateTime.now(),
       });
     } catch (e) {
       error = 'クレーム／要望情報の削除に失敗しました';
